Separate test code from app server
